massive code clean up
git-svn-id: https://babelfish.arc.nasa.gov/svn/stereopipeline/sandbox/lima@1481 1395e39a-8119-0410-9602-abf82d98ffd9

+++ display.cc
@@ -212,7 +212,7 @@ void MakeGrid(vector<vector<LOLAShot> >trackPts, int numVerPts, int numHorPts, s
   }
   //fill the DEM
 
-  printf("numTracks = %d\n", trackIndices.size());
+  printf("numTracks = %d\n", (int)(trackIndices.size()));
   for (int k = 0; k < trackIndices.size();k++){
     int trackIndex = trackIndices[k];
     printf("trackIndex = %d\n", trackIndex);
